app/data_collection/requirements.txt
python-twitter
neo4j
pattern
newsapi-python
azure-cognitiveservices-search-newssearch